Reflow all the language stack `README-content.md` files to 80 columns and incorporate Fred's changes
This also includes Fred's updates to the `README-short.txt` files, and a bunch of very minor changes for consistency (using "PHP" instead of "php", ending sentences that describe and lead into a code block with a ":", using "Go" instead of "Golang" and "Hy" instead of "Hylang" within prose, not using periods at the end of headlines, etc).

A really fun one included here is that Java mentioned both using GNU Make inside the container (and `java` doesn't include `make` presently), and some of the prose mentioned `go build` after we ran `javac`, which was extra neat.