Add a script for pushing descriptions trivially via vimdiff
This is inspired by Gentoo's `dispatch-conf`, which I've got configured to use `vimdiff` on all my Gentoo boxen, and it's just about the greatest thing ever.

This has a nasty little `monkey_patch` to Mojo's `CookieJar` that needs more digging to figure out what's going on there, but otherwise it's pretty solid.