Tweak the IRC language slightly by making it a separate paragraph
This is especially useful for READMEs where %%MAILING_LIST%% is actually plugged in, making a three-part list that we can't easily add an Oxford comma to.

Also, this allows us to explain that not all the image maintainers will necessarily be in that IRC channel ("You can also reach _many_ ...").