Update the postgresql README to show automatic daemon exit
The `docker-entrypoint.sh` script always starts the postgresql daemon regardless. I wanted to script  the process of creating a local user as described, so rather than having to hit CTRL+C you can force an exit by passing the `--help` argument so this it self exist on completion.
I remember in passing seeing people asking this question in somewhere else so hopefuly this README update will help in the future.