nginx: ensure best practice when using envsubst
Background: `envsubst` replaces all references that look like references
to variables including those that do not have a corresponding environment
variable - it just replaces those with an empty string. Since it's very
common for nginx.conf to have references like $uri or $host, which are
not definately not to be replaced by environment variables, there is a
need to limit the scope of `envsubsts` by specifying each variable name.

https://stackoverflow.com/questions/46473266/nginx-on-kubernetes-docker-doing-infinite-redirect-when-generating-conf

Also, using `exec nginx` is better, since we can avoid an extra shell
being run and can also make sure that nginx receives the signals
correctly.