Updating postgres notes/example
The example init script in this documentation fails when the `$POSTGRES_USER` and `$POSTGRES_DB` values mismatch.  By default, `psql` attempts to set the database in use to the same name as the logging in `$POSTGRES_USER`, so without setting the `-d` flag users will encounter an error similar to: psql: FATAL:  database "root" does not exist when they don't match.