Remove SELinux specific notes
This really isn't our place to document -- this is a generic Docker issue for users of SELinux.

The official upstream documentation on the topic can be found at https://docs.docker.com/storage/bind-mounts/#configure-the-selinux-label.

There's also some good information to be found in http://jaormx.github.io/2018/selinux-and-docker-notes/ and https://www.projectatomic.io/blog/2015/06/using-volumes-with-docker-can-cause-problems-with-selinux/.