Odoo: add missing database parameter to docker run commands
As of Odoo 12, the database parameter needs to be explicitly provided, otherwise
the database will not be initialized. Our example `docker` run commands need to be
updated to reflect this, otherwise users risk meeting an error page if they
connect to the Odoo web interface, as the implicit database "odoo" still needs
to be initialized.

For older versions this extra parameter will be accepted as well without causing
any issue.