Add warning about using PG env vars
I discovered that the image will not successfully fully initialize and will crash if the environment variables PGHOST, PGDATABASE, PGUSER and PGPASSWORD are included. I thought I'd use the same configmap for client and server in a k8s configuration so that changing, say, a password, would occur in just one place, but I discovered that the Postgres image hangs and then crashes when these environment variables are set. I removed them and created a separate client configmap. All is working OK again.