Add comment about database host name
This will make it easier to set up the mediawiki for users not intimately familiar with Docker Stacks conventions for service host naming. I was lucky to find the answer to this problem on https://stackoverflow.com/questions/57312109/mediawiki-docker-official-image-connection-refused-by-mysql, but having it in the docs would make the setup much more smooth.