Fix container path for cache dir
The code block contains the paths that should be mapped,
`/var/cache/nginx` and `/var/run` for readonly cases.
However, the description immediately above the code block
also listed two paths. The first, `/var/cache`, would require
additional dirs to already exist within the dir. For example,
the application will fail to start if the host didn't create
an `nginx` dir prior to starting the container.

This change updates the path in the description so it matches
the path mapped in the code block and no dirs need to exist
before starting the container.