ubuntu: Update Git references to Launchpad
Ubuntu base images rootfs tarballs are now fetched from Launchpad Git
repository. The location https://partner-images.canonical.com/oci/ is no
longer used. Update links in the documentation to refer to the
https://git.launchpad.net/cloud-images/+oci/ubuntu-base Git repository
URL instead.

The Launchpad Git repository URL is also set in github-repo file. This
shouldn't matter because the fact that an URL is a GitHub URL is used
only to fetch badges from open-source CI/CD services, and in case the
URL is not a GitHub URL, the HTTP request will just fail, and no badge
will be included.