Trim descriptions based on bytes, not characters
It turns out that the Hub description length limit is 25k *bytes*, not characters.  We've gotten lucky that our few unicode glyphs have landed so far such that we've never run into a problem with this before now, but we're currently hitting this while trying to push the `amd64/nextcloud` description in that the trimmed version ends up being exactly 25,003 bytes thanks to the unicode glyphs throwing off our trimming calculations (and getting really [un]lucky with the exact positioning of the markdown such that our trimming doesn't walk back far enough).

This fixes that by calculating and trimming based on bytes, accounting for cases of unicode glyphs directly spanning the 25k byte boundary.

After applying this fix, the same breaking `amd64/nextcloud` trimmed description is 24,725 bytes instead.